move node loading into its own class
